pyproject.toml
[project]
name = "news-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.55.0",
    "reportlab>=4.4.6",
    "requests>=2.32.5",
    "streamlit>=1.52.1",
]
